Fixed issue with constraint not being at table level
